push all i have

+++ movie/src/main/java/com/revature/movie/web/controllers/DeleteUser.java
@@ -2,6 +2,7 @@ package com.revature.movie.web.controllers;
 
 
 import com.revature.movie.model.User;
+import com.revature.movie.model.UserRole;
 import com.revature.movie.services.UserService;
 import com.revature.movie.web.dtos.UnblockCreds;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -26,7 +27,13 @@ public class DeleteUser {
 
     @PostMapping(produces= MediaType.APPLICATION_JSON_VALUE, consumes=MediaType.APPLICATION_JSON_VALUE)
     public void delete(@RequestBody UnblockCreds username, HttpServletResponse response) {
-        userService.deleteUser(username.getUsername());
+
+        if (username.getRole() == UserRole.ADMIN){
+            userService.deleteUser(username.getUsername());
+        }else {
+            response.setStatus(HttpServletResponse.SC_NON_AUTHORITATIVE_INFORMATION);//203
+        }
+
 
     }
 }
